+++ ff/Main.hs
@@ -32,8 +32,8 @@ import           Text.PrettyPrint.Mainland.Class (Pretty, ppr)
 
 import           FF (cmdDeleteContact, cmdDeleteNote, cmdDone, cmdEdit,
                      cmdNewContact, cmdNewNote, cmdPostpone, cmdSearch,
-                     cmdUnarchive, getContactSamples, getSamples, getUtcToday,
-                     getWikiSamples, updateTrackedNotes)
+                     cmdUnarchive, getContactSamples, getNoteSamples,
+                     getUtcToday, getWikiSamples, updateTrackedNotes)
 import           FF.Config (Config (..), ConfigUI (..), appName, loadConfig,
                             printConfig, saveConfig)
 import           FF.Github (getIssueViews, getOpenIssueSamples)
@@ -120,9 +120,9 @@ runCmdAction h ui cmd brief = do
     today <- getUtcToday
     case cmd of
         CmdAgenda mlimit -> do
-            nvs <- getSamples ui mlimit today
+            nvs <- getNoteSamples ui mlimit today
             pprint $ UI.prettySamplesBySections brief nvs
-        CmdContact contact -> cmdContact contact
+        CmdContact contact -> cmdContact brief contact
         CmdDelete noteId -> do
             nv <- cmdDeleteNote noteId
             pprint $ withHeader "deleted:" $ UI.noteViewFull nv
@@ -139,8 +139,9 @@ runCmdAction h ui cmd brief = do
             nv <- cmdPostpone noteId
             pprint $ withHeader "postponed:" $ UI.noteViewFull nv
         CmdSearch Search {..} -> do
-            nvs <- cmdSearch searchText searchLimit today
-            pprint $ UI.prettySamplesBySections brief nvs
+            (notes, wiki, contacts) <- cmdSearch searchText searchLimit today
+            pprint $ UI.prettyNotesWikiContacts
+                brief notes wiki contacts searchTasks searchWiki searchContacts
         CmdServe -> cmdServe h ui
         CmdTrack track ->
             cmdTrack track today brief
@@ -152,7 +153,7 @@ runCmdAction h ui cmd brief = do
             liftIO $ putStrLn "upgraded"
         CmdWiki mlimit -> do
             nvs <- getWikiSamples ui mlimit today
-            pprint $ UI.prettySamplesBySections brief nvs
+            pprint $ UI.prettyWikiSamplesOmitted brief nvs
 
 cmdTrack :: Track -> Day -> Bool -> Storage ()
 cmdTrack Track {..} today brief =
@@ -178,8 +179,8 @@ cmdTrack Track {..} today brief =
                 exitFailure
             Right issues -> pure issues
 
-cmdContact :: Maybe Contact -> Storage ()
-cmdContact = \case
+cmdContact :: Bool -> Maybe Contact -> Storage ()
+cmdContact brief = \case
     Just (Add name) -> do
         cv <- cmdNewContact name
         pprint $ withHeader "added:" $ UI.contactViewFull cv
@@ -188,7 +189,7 @@ cmdContact = \case
         pprint $ withHeader "deleted:" $ UI.contactViewFull cv
     Nothing -> do
         cvs <- getContactSamples
-        pprint $ UI.prettyContactSamplesOmitted cvs
+        pprint $ UI.prettyContactSamplesOmitted brief cvs
 
 -- Template taken from stack:
 -- "Version 1.7.1, Git revision 681c800873816c022739ca7ed14755e8 (5807 commits)"